# Convert Block and Type from interface to abstract class