--- a/presto-main/src/main/java/com/facebook/presto/operator/aggregation/AccumulatorCompiler.java
+++ b/presto-main/src/main/java/com/facebook/presto/operator/aggregation/AccumulatorCompiler.java
@@ -548,35 +548,35 @@ public class AccumulatorCompiler
                     .append(constantType(callSiteBinder, sqlType))
                     .append(getBlockBytecode)
                     .append(position)
-                    .invokeInterface(Type.class, "getLong", long.class, Block.class, int.class);
+                    .invokeVirtual(Type.class, "getLong", long.class, Block.class, int.class);
         }
         else if (parameter == double.class) {
             block.comment("%s.getDouble(block, position)", sqlType.getTypeSignature())
                     .append(constantType(callSiteBinder, sqlType))
                     .append(getBlockBytecode)
                     .append(position)
-                    .invokeInterface(Type.class, "getDouble", double.class, Block.class, int.class);
+                    .invokeVirtual(Type.class, "getDouble", double.class, Block.class, int.class);
         }
         else if (parameter == boolean.class) {
             block.comment("%s.getBoolean(block, position)", sqlType.getTypeSignature())
                     .append(constantType(callSiteBinder, sqlType))
                     .append(getBlockBytecode)
                     .append(position)
-                    .invokeInterface(Type.class, "getBoolean", boolean.class, Block.class, int.class);
+                    .invokeVirtual(Type.class, "getBoolean", boolean.class, Block.class, int.class);
         }
         else if (parameter == Slice.class) {
             block.comment("%s.getSlice(block, position)", sqlType.getTypeSignature())
                     .append(constantType(callSiteBinder, sqlType))
                     .append(getBlockBytecode)
                     .append(position)
-                    .invokeInterface(Type.class, "getSlice", Slice.class, Block.class, int.class);
+                    .invokeVirtual(Type.class, "getSlice", Slice.class, Block.class, int.class);
         }
         else {
             block.comment("%s.getObject(block, position)", sqlType.getTypeSignature())
                     .append(constantType(callSiteBinder, sqlType))
                     .append(getBlockBytecode)
                     .append(position)
-                    .invokeInterface(Type.class, "getObject", Object.class, Block.class, int.class);
+                    .invokeVirtual(Type.class, "getObject", Object.class, Block.class, int.class);
         }
     }