--- a/presto-spi/src/main/java/com/facebook/presto/spi/block/AbstractSingleRowBlock.java
+++ b/presto-spi/src/main/java/com/facebook/presto/spi/block/AbstractSingleRowBlock.java
@@ -17,7 +17,7 @@ package com.facebook.presto.spi.block;
 import io.airlift.slice.Slice;
 
 public abstract class AbstractSingleRowBlock
-        implements Block
+        extends BlockBuilder
 {
     protected final int rowIndex;