--- a/presto-spi/src/main/java/com/facebook/presto/spi/block/BlockBuilder.java
+++ b/presto-spi/src/main/java/com/facebook/presto/spi/block/BlockBuilder.java
@@ -15,13 +15,13 @@ package com.facebook.presto.spi.block;
 
 import io.airlift.slice.Slice;
 
-public interface BlockBuilder
+public abstract class BlockBuilder
         extends Block
 {
     /**
      * Write a byte to the current entry;
      */
-    default BlockBuilder writeByte(int value)
+    public BlockBuilder writeByte(int value)
     {
         throw new UnsupportedOperationException(getClass().getName());
     }
@@ -29,7 +29,7 @@ public interface BlockBuilder
     /**
      * Write a short to the current entry;
      */
-    default BlockBuilder writeShort(int value)
+    public BlockBuilder writeShort(int value)
     {
         throw new UnsupportedOperationException(getClass().getName());
     }
@@ -37,7 +37,7 @@ public interface BlockBuilder
     /**
      * Write a int to the current entry;
      */
-    default BlockBuilder writeInt(int value)
+    public BlockBuilder writeInt(int value)
     {
         throw new UnsupportedOperationException(getClass().getName());
     }
@@ -45,7 +45,7 @@ public interface BlockBuilder
     /**
      * Write a long to the current entry;
      */
-    default BlockBuilder writeLong(long value)
+    public BlockBuilder writeLong(long value)
     {
         throw new UnsupportedOperationException(getClass().getName());
     }
@@ -53,7 +53,7 @@ public interface BlockBuilder
     /**
      * Write a byte sequences to the current entry;
      */
-    default BlockBuilder writeBytes(Slice source, int sourceIndex, int length)
+    public BlockBuilder writeBytes(Slice source, int sourceIndex, int length)
     {
         throw new UnsupportedOperationException(getClass().getName());
     }
@@ -62,7 +62,7 @@ public interface BlockBuilder
      * Return a writer to the current entry. The caller can operate on the returned caller to incrementally build the object. This is generally more efficient than
      * building the object elsewhere and call writeObject afterwards because a large chunk of memory could potentially be unnecessarily copied in this process.
      */
-    default BlockBuilder beginBlockEntry()
+    public BlockBuilder beginBlockEntry()
     {
         throw new UnsupportedOperationException(getClass().getName());
     }
@@ -71,7 +71,7 @@ public interface BlockBuilder
      * Create a new block from the current materialized block by keeping the same elements
      * only with respect to {@code visiblePositions}.
      */
-    default Block getPositions(int[] visiblePositions, int offset, int length)
+    public Block getPositions(int[] visiblePositions, int offset, int length)
     {
         return build().getPositions(visiblePositions, offset, length);
     }
@@ -79,17 +79,23 @@ public interface BlockBuilder
     /**
      * Write a byte to the current entry;
      */
-    BlockBuilder closeEntry();
+    public BlockBuilder closeEntry()
+    {
+        throw new UnsupportedOperationException(getClass().getName());
+    }
 
     /**
      * Appends a null value to the block.
      */
-    BlockBuilder appendNull();
+    public BlockBuilder appendNull()
+    {
+        throw new UnsupportedOperationException(getClass().getName());
+    }
 
     /**
      * Append a struct to the block and close the entry.
      */
-    default BlockBuilder appendStructure(Block value)
+    public BlockBuilder appendStructure(Block value)
     {
         throw new UnsupportedOperationException(getClass().getName());
     }
@@ -98,7 +104,7 @@ public interface BlockBuilder
      * Do not use this interface outside block package.
      * Instead, use Block.writePositionTo(BlockBuilder, position)
      */
-    default BlockBuilder appendStructureInternal(Block block, int position)
+    public BlockBuilder appendStructureInternal(Block block, int position)
     {
         throw new UnsupportedOperationException(getClass().getName());
     }
@@ -106,10 +112,16 @@ public interface BlockBuilder
     /**
      * Builds the block. This method can be called multiple times.
      */
-    Block build();
+    public Block build()
+    {
+        throw new UnsupportedOperationException(getClass().getName());
+    }
 
     /**
      * Creates a new block builder of the same type based on the current usage statistics of this block builder.
      */
-    BlockBuilder newBlockBuilderLike(BlockBuilderStatus blockBuilderStatus);
+    public BlockBuilder newBlockBuilderLike(BlockBuilderStatus blockBuilderStatus)
+    {
+        throw new UnsupportedOperationException(getClass().getName());
+    }
 }