Based on @oerling findings.

Interface method dispatch is more expensive than the virtual method dispatch, as it requires additional level of indirection.

Interface / Virtual method dispatch can be optimized by the JVM in call site is monomorphic / bimorphic. Unfortunately our code generation is not ideal, and somethimes JVM ends up executing megamorphic invokations without optimizations by doing ITable traversal + VTable lookups. 

Optimizing our code generation to have all the Block/Type methods invocations as monomorphic is definitely the best long term solution. But for the short term solution we can cut the megamorphic dispatch overhead by the factor of 2 by changing the `Block` and `Type` to be an abstract class instead of an interface. 

Below are the results of running `BenchmarkPageProcessor` before and after this change

```
BEFORE

Benchmark                          Mode  Cnt     Score    Error  Units
BenchmarkPageProcessor.compiled   thrpt   10  6096.074 ± 57.174  ops/s
BenchmarkPageProcessor.handCoded  thrpt   10  6861.199 ± 47.061  ops/s
```

```
AFTER

Benchmark                          Mode  Cnt     Score    Error  Units
BenchmarkPageProcessor.compiled   thrpt   10  6246.534 ± 51.951  ops/s
BenchmarkPageProcessor.handCoded  thrpt   10  6757.921 ± 63.097  ops/s
```

The benchmarks shows an improvement in `~3%`. However the  `BenchmarkPageProcessor` runs comparisons on `VARCHAR`, that takes most of the CPU time. After removing the `VARCHAR` comparisons, but leaving the other three `DOUBLE` comparisons, the benchmark shows `~20%` improvement.


```
BEFORE, WITHOUT VARCHAR

Benchmark                          Mode  Cnt      Score     Error  Units
BenchmarkPageProcessor.compiled   thrpt   10   8981.628 ±  51.715  ops/s
BenchmarkPageProcessor.handCoded  thrpt   10  14339.662 ± 983.400  ops/s
```

```
AFTER, WITHOUT VARCHAR

Benchmark                          Mode  Cnt      Score     Error  Units
BenchmarkPageProcessor.compiled   thrpt   10  11073.608 ± 232.180  ops/s
BenchmarkPageProcessor.handCoded  thrpt   10  14335.648 ± 621.870  ops/s
```

As a follow up we should ensure that `InputReferenceCompiler` generates exact method signatures for the `Type` and `Block`, instead of generic ones:

![selection_038](https://user-images.githubusercontent.com/5570988/46634790-123a4200-cb20-11e8-90d8-37b24991d673.png)