@oerling 2018-10-08
For a query like:

 

 

select count (*), sum (l.extendedprice * (1 - l.discount) ) from raptor.tpch.lineitem l, raptor.tpch.partsupp p where l.partkey = p.partkey and l.suppkey = p.suppkey and p.availqty < 1000;

 

 

The itable share of profile is:

 

11890114962    2.41%     1189  itable stub

  8360191653    1.70%      836  itable stub

  3130407696    0.64%      313  itable stub

  2230090074    0.45%      223  itable stub

   521026635    0.11%       52  itable stub

   210071157    0.04%       21  itable stub

   169994970    0.03%       17  itable stub

   130206598    0.03%       13  itable stub

 

 

So, if vtable is 2x faster than itable then the difference will not be something the end user will notice. On the other hand there does not seem to be any specific reason why Block should be an interface.

 

 

In the end in going for speed the point is to dispatch once per  column  and not once  per column*row, in which case the itable vs vtable difference will not be visible.  In principle the JVM should be able to do this via monomorphic inlining. In practice this does not seem to happen. 

 

 

 

 

 

 

 

From: Dain Sundstrom <notifications@github.com> 
Sent: Monday, October 8, 2018 2:54 PM
To: prestodb/presto <presto@noreply.github.com>
Cc: oerling <erling@xs4all.nl>; Mention <mention@noreply.github.com>
Subject: Re: [prestodb/presto] Convert Block and Type from interface to abstract class (#11663)

 

@dain requested changes on this pull request.

I don't think we should not be making abstraction changes like this unless there is proof that this effects real world queries, the result would be significant, and there are no other options available.

For the first part, it looks like this does not effect real world queries since as you mentioned "After removing the VARCHAR comparisons, but leaving the other three DOUBLE comparisons, the benchmark shows ~20% improvement.", but queries with only simple math are rare, and are already are super fast. Which brings us to the second part. This means they represent a tiny portion of global CPU, so I would expect this impact of this optimization to be negligible in a production cluster. Along that path, at FB what % of CPU time is spent in code that would benefit from this change. Assuming, you can pass the threshold for the first two parts, I'd dig into why you are getting metamorphic call sites, for the third part. In the past we avoided these by good code generation, and tricks like reloading functions for each query to avoid profile pollution from other queries. If that doesn't, work then we should look into alternative solutions... maybe something like using invoke dynamic instead of invoke interface.

—
You are receiving this because you were mentioned.
Reply to this email directly, view it on GitHub <https://github.com/prestodb/presto/pull/11663#pullrequestreview-162652572> , or mute the thread <https://github.com/notifications/unsubscribe-auth/Ap73z_Nc2yRp1ZcXBErHgqVyJ7dYFv_iks5ui8mVgaJpZM4XNtm8> .