@arhimondr 2018-10-15
I don't know what was wrong. But i made the ASM dumps, and it looks like the inlining works really great for filter function. So there is no real difference between interface and abstract class. I'm closing it for now.